fix: moves resolvesTargetCompanyTrait to kanvas
